+++ src/Picker/Classes/WindowControllers/PickerWindowController.swift
@@ -1,7 +1,7 @@
 import AppKit
 
 /// Window controller for the floating picker window
-class PickerWindowController: NSWindowController {
+class PickerWindowController: NSWindowController, NSWindowDelegate {
     // MARK: - Properties
 
     private let pickerViewController: PickerViewController
@@ -26,10 +26,24 @@ class PickerWindowController: NSWindowController {
         window.level = .floating
 
         super.init(window: window)
+        window.delegate = self
     }
 
     @available(*, unavailable)
     required init?(coder _: NSCoder) {
         fatalError("init(coder:) has not been implemented")
+    }
+
+    // MARK: - Window Lifecycle
+
+    override func showWindow(_ sender: Any?) {
+        super.showWindow(sender)
+        ColorPicker.shared.previewDidBecomeVisible()
+    }
+
+    // MARK: - NSWindowDelegate
+
+    func windowWillClose(_: Notification) {
+        ColorPicker.shared.previewDidBecomeHidden()
     }
 }